updated backend datetime handling

+++ backend/resources/transaction.py
@@ -4,34 +4,37 @@ from models.transaction import TransactionModel
 import datetime
 
 
+parser = reqparse.RequestParser()
+parser.add_argument(
+    'date_time',
+    type=str,
+    required=True,
+    help="Transaction date and time are required"
+)
+parser.add_argument(
+    'account_from',
+    type=str,
+    required=False
+)
+parser.add_argument(
+    'account_to',
+    type=str,
+    required=False
+)
+parser.add_argument(
+    'cost',
+    type=float,
+    required=True,
+    help="Transaction value is required"
+)
+
+
 class Transaction(Resource):
-    parser = reqparse.RequestParser()
-    parser.add_argument(
-        'date_time',
-        type=datetime,
-        required=True,
-        help="Transaction date and time are required"
-    )
-    parser.add_argument(
-        'account_from',
-        type=str,
-        required=False
-    )
-    parser.add_argument(
-        'account_to',
-        type=str,
-        required=False
-    )
-    parser.add_argument(
-        'cost',
-        type=float,
-        required=True,
-        help="Transaction value is required"
-    )
 
     @jwt_required
     def post(self, user_id):
-        data = self.parser.parse_args()
+        data = parser.parse_args()
+        data.date_time = datetime.datetime.strptime(parser.parse_args().date_time, '%Y-%m-%dT%H:%M:%S')
 
         transaction = TransactionModel(user_id, **data)
 
@@ -42,35 +45,54 @@ class Transaction(Resource):
 
         return transaction.json(), 201
 
+class TransactionById(Resource):
     @jwt_required
-    def get(self, _id):
-        transaction = TransactionModel.find_by_id(_id)
+    def get(self, user_id, _id):
+        if user_id == get_jwt_identity():
+            transaction = TransactionModel.find_by_id(_id)
 
-        if transaction & transaction.user_id == get_jwt_identity():
-            return transaction.json()
+            if transaction & transaction.user_id == get_jwt_identity():
+                return transaction.json()
 
-        return {'message': 'Transaction not found'}
+            return {'message': 'Transaction not found'}
 
-    @jwt_required
-    def delete(self, _id):
-        transaction = TransactionModel.find_by_id(_id)
-
-        if transaction & transaction.user_id == get_jwt_identity():
-            transaction.delete_from_db()
-            return {'message': f'Transaction with id:{_id} deleted'}
-
-        return {'message': 'Transaction not found'}
+        return {'message': 'You can only fetch your own transactions'}
 
     @jwt_required
-    def put(self, _id):
-        data = self.parser.parse_args()
+    def delete(self, user_id, _id):
+        if user_id == get_jwt_identity():
+            transaction = TransactionModel.find_by_id(_id)
 
-        transaction = TransactionModel.find_by_id(_id)
+            if transaction & transaction.user_id == get_jwt_identity():
+                transaction.delete_from_db()
+                return {'message': f'Transaction with id:{_id} deleted'}
 
-        if transaction:
-            for key in data:
-                transaction[key] = data[key]
-            transaction.save_to_db() 
-            return transaction.json()
+            return {'message': 'Transaction not found'}
 
-        return {'message': 'Transaction not found'}
+        return {'message': 'You can only delete your own transactions'}
+
+    @jwt_required
+    def put(self, user_id, _id):
+        if user_id == get_jwt_identity():
+            data = parser.parse_args()
+            transaction = TransactionModel.find_by_id(_id)
+
+            if transaction:
+                for key in data:
+                    transaction[key] = data[key]
+                transaction.save_to_db()
+                return transaction.json()
+
+            return {'message': 'Transaction not found'}
+
+        return {'message': 'You can only edit your own transactions'}
+
+
+class TransactionList(Resource):
+    @jwt_required
+    def get(self, user_id):
+        transactions = {transaction.json()
+                        for transaction in TransactionModel.find_all(user_id)}
+        if transactions:
+            return transactions, 200
+        return {'message': 'no transactions available'}, 200
